responsive design and color changes
